feat: add weather units to api calls

+++ app/src/main/java/com/gzaber/weatherapp/data/repository/weather/model/CurrentWeather.kt
@@ -4,6 +4,6 @@ data class CurrentWeather(
     val weatherCode: Int,
     val temperature: CurrentWeatherParameter<Double>,
     val humidity: CurrentWeatherParameter<Int>,
-    val rain: CurrentWeatherParameter<Int>,
+    val rain: CurrentWeatherParameter<Double>,
     val windSpeed: CurrentWeatherParameter<Double>
 )
